ai-service/food_api/requirements.txt
fastapi
uvicorn
onnxruntime
numpy
opencv-python
python-multipart
httpx
Pillow
jinja2
aiofiles